Add: config.py 추가, SupabaseConfig 추가
