+++ service/src/supabase_handler.py
@@ -3,7 +3,6 @@ import sys
 import logging
 import yaml
 
-# from dotenv import load_dotenv
 from supabase import create_client, Client
